# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * One for All – voice-first personal capture API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EntryCaptureType = typeof EntryCaptureType[keyof typeof EntryCaptureType];


export const EntryCaptureType = {
  voice: 'voice',
  text: 'text',
} as const;

export type EntryCategory = typeof EntryCategory[keyof typeof EntryCategory];


export const EntryCategory = {
  inbox: 'inbox',
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;

/**
 * @nullable
 */
export type EntrySuggestedCategory = typeof EntrySuggestedCategory[keyof typeof EntrySuggestedCategory] | null;


export const EntrySuggestedCategory = {
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;

export interface Entry {
  id: number;
  content: string;
  captureType: EntryCaptureType;
  category: EntryCategory;
  isTaskDone: boolean;
  /** @nullable */
  suggestedCategory: EntrySuggestedCategory;
  createdAt: string;
  updatedAt: string;
}

export interface Person {
  id: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  /**
     * Optional short label to distinguish same-named people, e.g. "Studentina", "climbing gym"
     * @nullable
     */
  descriptor?: string | null;
  createdAt: string;
}

export type EntryDetail = Entry & {
  people: Person[];
};

export type EntryInputCaptureType = typeof EntryInputCaptureType[keyof typeof EntryInputCaptureType];


export const EntryInputCaptureType = {
  voice: 'voice',
  text: 'text',
} as const;

export type EntryInputCategory = typeof EntryInputCategory[keyof typeof EntryInputCategory];


export const EntryInputCategory = {
  inbox: 'inbox',
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;

export type EntryInputSuggestedCategory = typeof EntryInputSuggestedCategory[keyof typeof EntryInputSuggestedCategory];


export const EntryInputSuggestedCategory = {
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;

export interface EntryInput {
  /** @minLength 1 */
  content: string;
  captureType: EntryInputCaptureType;
  category?: EntryInputCategory;
  suggestedCategory?: EntryInputSuggestedCategory;
}

export type EntryUpdateCategory = typeof EntryUpdateCategory[keyof typeof EntryUpdateCategory];


export const EntryUpdateCategory = {
  inbox: 'inbox',
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;

export type EntryUpdateSuggestedCategory = typeof EntryUpdateSuggestedCategory[keyof typeof EntryUpdateSuggestedCategory];


export const EntryUpdateSuggestedCategory = {
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;

export interface EntryUpdate {
  /** @minLength 1 */
  content?: string;
  category?: EntryUpdateCategory;
  isTaskDone?: boolean;
  suggestedCategory?: EntryUpdateSuggestedCategory;
}

export interface EntryStats {
  inbox: number;
  journal: number;
  task: number;
  idea: number;
  log: number;
  total: number;
}

export interface LinkPersonInput {
  personId: number;
}

export type PersonDetail = Person & {
  entries: Entry[];
};

export interface PersonInput {
  /** @minLength 1 */
  name: string;
  notes?: string;
  /** Optional short label to distinguish same-named people */
  descriptor?: string;
}

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  notes?: string;
  /** Optional short label to distinguish same-named people */
  descriptor?: string;
}

export type ListEntriesParams = {
category?: ListEntriesCategory;
};

export type ListEntriesCategory = typeof ListEntriesCategory[keyof typeof ListEntriesCategory];


export const ListEntriesCategory = {
  inbox: 'inbox',
  journal: 'journal',
  task: 'task',
  idea: 'idea',
  log: 'log',
} as const;