# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: One for All – voice-first personal capture API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people:
    get:
      operationId: listPeople
      tags: [people]
      summary: List all people
      responses:
        "200":
          description: List of people
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
    post:
      operationId: createPerson
      tags: [people]
      summary: Create a new person record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonInput"
      responses:
        "201":
          description: Created person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /people/{id}:
    get:
      operationId: getPerson
      tags: [people]
      summary: Get a person with their linked entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Person detail with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePerson
      tags: [people]
      summary: Update a person's name or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonUpdate"
      responses:
        "200":
          description: Updated person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
    delete:
      operationId: deletePerson
      tags: [people]
      summary: Delete a person record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Entry ────────────────────────────────────────────────────────────────
    Entry:
      type: object
      required: [id, content, captureType, category, isTaskDone, suggestedCategory, createdAt, updatedAt]
      properties:
        id:
          type: integer
        content:
          type: string
        captureType:
          type: string
          enum: [voice, text]
        category:
          type: string
          enum: [inbox, journal, task, idea, log]
        isTaskDone:
          type: boolean
        suggestedCategory:
          type: ["string", "null"]
          enum: [journal, task, idea, log, null]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Person:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        descriptor:
          type: ["string", "null"]
          description: Optional short label to distinguish same-named people, e.g. "Studentina", "climbing gym"
        createdAt:
          type: string
          format: date-time
    PersonDetail:
      allOf:
        - $ref: "#/components/schemas/Person"
        - type: object
          required: [entries]
          properties:
            entries:
              type: array
              items:
                $ref: "#/components/schemas/Entry"
    PersonInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        notes:
          type: string
        descriptor:
          type: string
          description: Optional short label to distinguish same-named people
    PersonUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        notes:
          type: string
        descriptor:
          type: string
          description: Optional short label to distinguish same-named people