# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * One for All – voice-first personal capture API
 * OpenAPI spec version: 0.1.0
 */

export interface Person {
  id: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  /**
     * Optional short label to distinguish same-named people, e.g. "Studentina", "climbing gym"
     * @nullable
     */
  descriptor?: string | null;
  createdAt: Date;
}