# lib/api-zod/src/generated/types/personUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * One for All – voice-first personal capture API
 * OpenAPI spec version: 0.1.0
 */

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  notes?: string;
  /** Optional short label to distinguish same-named people */
  descriptor?: string;
}